Refactor and comment code
