+++ app/client/src/layoutSystems/anvil/utils/layouts/update/mainCanvasLayoutUtils.ts
@@ -13,7 +13,7 @@ import { call } from "redux-saga/effects";
 import { severTiesFromParents, transformMovedWidgets } from "./moveUtils";
 import { anvilWidgets } from "widgets/anvil/constants";
 import { MAIN_CONTAINER_WIDGET_ID } from "constants/WidgetConstants";
-import { addNewAnvilWidgetToDSL } from "layoutSystems/anvil/integrations/sagas/anvilDraggingSagas";
+import { addNewAnvilWidgetToDSL } from "layoutSystems/anvil/integrations/sagas/anvilWidgetAdditionSagas/helpers";
 
 /**
  * This function adds a detached widget to the main canvas.
