+++ app/client/src/layoutSystems/anvil/utils/sectionOperationUtils.ts
@@ -10,11 +10,11 @@ import type {
 import { call, select } from "redux-saga/effects";
 import { getWidgets } from "sagas/selectors";
 import { generateReactKey } from "utils/generators";
-import { addNewChildToDSL } from "../integrations/sagas/anvilDraggingSagas";
 import type BaseLayoutComponent from "../layoutComponents/BaseLayoutComponent";
 import LayoutFactory from "../layoutComponents/LayoutFactory";
 import { defaultHighlightRenderInfo } from "../utils/constants";
 import { anvilWidgets } from "widgets/anvil/constants";
+import { getUpdatedListOfWidgetsAfterAddingNewWidget } from "../integrations/sagas/anvilWidgetAdditionSagas";
 
 /**
  * Function to get the highlight information for the last column of a Section Widget
@@ -173,7 +173,7 @@ export function* addNewZonesToSection(
     };
     const highlight = getSectionLastColumnHighlight(sectionWidget);
     updatedWidgets = yield call(
-      addNewChildToDSL,
+      getUpdatedListOfWidgetsAfterAddingNewWidget,
       highlight,
       newWidget,
       false,
