+++ app/client/src/widgets/anvil/ZoneWidget/widget/config/defaultConfig.ts
@@ -30,18 +30,36 @@ export const defaultConfig: WidgetDefaultProps = {
         fn: (
           widget: FlattenedWidgetProps,
           widgets: CanvasWidgetsReduxState,
-          parent: FlattenedWidgetProps,
-          layoutSystemType: LayoutSystemTypes,
+          parent: FlattenedWidgetProps, // Why does this exist, when we have all the widgets?
+          layoutSystemType: LayoutSystemTypes, // All widgets are new in Anvil, however, it may be needed for Auto Layout
         ) => {
           if (layoutSystemType !== LayoutSystemTypes.ANVIL) return [];
 
           const layout: LayoutProps[] = zonePreset();
 
-          return getWidgetBluePrintUpdates({
+          const updates = getWidgetBluePrintUpdates({
             [widget.widgetId]: {
               layout,
             },
           });
+
+          // In a modal widget, the zones don't have borders and elevation
+          // by default. We go up the hierarchy to find any Modal Widget
+          // If it exists, we remove the elevated background for the zone
+          let parentId = widget.parentId;
+          while (parentId) {
+            if (widgets[parentId].type === "WDS_MODAL_WIDGET") {
+              updates.push({
+                widgetId: widget.widgetId,
+                propertyName: "elevatedBackground",
+                propertyValue: false,
+              });
+              break;
+            }
+            parentId = widgets[parentId].parentId;
+          }
+
+          return updates;
         },
       },
     ],
