+++ app/client/src/widgets/wds/WDSModalWidget/config/defaultsConfig.ts
@@ -34,7 +34,6 @@ export const defaultsConfig = {
           layoutSystemType: LayoutSystemTypes,
         ) => {
           if (layoutSystemType !== LayoutSystemTypes.ANVIL) return [];
-
           const layout: LayoutProps[] = modalPreset();
           return getWidgetBluePrintUpdates({
             [widget.widgetId]: {
